requirements.txt
pandas~=1.4.2
requests~=2.28.0
streamlit~=1.10.0
numpy~=1.22.4
psycopg2~=2.9.3
plotly.express~=0.4.1
greenlet==1.1.2
pathlib==1.0.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlalchemy==1.4.37
python-dotenv==0.15.0
